requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.9
django-paypal==1.0.0
djangorestframework==3.12.1
idna==2.10
mysqlclient==2.0.2
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.3.0
urllib3==1.26.2
cryptography==3.4.6
whitenoise==5.2.0
django-cloudinary-storage==0.2.3
django-heroku==0.3.1
gunicorn==20.0.4
django-filter
django-throttle-requests
openpyxl
django-google-analytics-app
